phase1/src/lattice_driver.py
from __future__ import annotations
import argparse
from pathlib import Path
import yaml

def main():
    ap = argparse.ArgumentParser()
    ap.add_argument("--config", required=True)
    ap.add_argument("--mode", choices=["amplitude","scaling"], required=True)
    ap.add_argument("--out", required=True)
    args = ap.parse_args()

    cfg = yaml.safe_load(Path(args.config).read_text())

    # Placeholder: next step will import and wrap your existing toy_universe_lattice code
    # and produce the canonical figures B and C.

    Path(args.out).parent.mkdir(parents=True, exist_ok=True)
    Path(args.out).write_text("TODO: implement lattice driver that generates this figure.\n")

if __name__ == "__main__":
    main()
